backend/requirements.txt
django
djangorestframework
django-environ
Pillow
django-extensions
psycopg2